+++ formsfeeder.core/src/main/java/com/_4point/aem/formsfeeder/core/datasource/DataSourceListDataSourceWrapper.java
@@ -2,11 +2,11 @@ package com._4point.aem.formsfeeder.core.datasource;
 
 public class DataSourceListDataSourceWrapper extends DataSourceWrapper {
 
-	private final ZipInputDataSourceWrapper zip;
+	private final ZipDataSourceWrapper zip;
 	
 	private DataSourceListDataSourceWrapper(DataSource dataSource) {
 		super(dataSource);
-		this.zip = ZipInputDataSourceWrapper.wrap(dataSource);
+		this.zip = ZipDataSourceWrapper.wrap(dataSource);
 	}
 	
 	public static DataSourceListDataSourceWrapper wrap(DataSource dataSource) {
