+++ formsfeeder.core/src/test/java/com/_4point/aem/formsfeeder/core/datasource/DataSourceListBuilderTest.java
@@ -147,7 +147,7 @@ class DataSourceListBuilderTest {
 	 * @param resultDsl
 	 */
 	void validateDslData(DataSource resultDs) {
-		assertTrue(resultDs instanceof ZipInputDataSourceWrapper, "Expected datasource to be instanceof ZipDataSource but it wasn't (" + resultDs.getClass().getName() + ").");
+		assertTrue(resultDs instanceof ZipDataSourceWrapper, "Expected datasource to be instanceof ZipDataSource but it wasn't (" + resultDs.getClass().getName() + ").");
 		// TODO: commented out until more of ZipDataSource is implemented 
 //		DataSourceList resultDsl = ((ZipDataSource) resultDs).asDataSourceList();
 //		assertEquals(dslData.size(), resultDsl.size());
